requirements.txt
pandas
dask[complete]
scikit-learn
xgboost
torch
fancyimpute
plotly
dash
python-pptx
watchdog
fastapi
uvicorn
stripe
coinbase-commerce
transformers
python-dotenv
supabase
joblib
matplotlib
scipy
pytest
